requirements.txt
mysql-connector-python
python-dotenv
Faker
